Update script to update examples
