+++ examples/models/package.json
@@ -8,7 +8,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "^4.0.2",
-    "upscaler": "0.8.27"
+    "upscaler": "0.8.17"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -31,5 +31,5 @@
       "last 1 safari version"
     ]
   },
-  "version": "0.8.27"
+  "version": "0.8.17"
 }
