+++ examples/patch-sizes/package.json
@@ -10,7 +10,7 @@
     "react-input-range": "^1.3.0",
     "react-scripts": "^4.0.3",
     "tensor-as-base64": "^0.1.1",
-    "upscaler": "0.8.27"
+    "upscaler": "0.8.17"
   },
   "scripts": {
     "start": "PORT=1234 react-scripts start"
@@ -30,5 +30,5 @@
       "last 1 safari version"
     ]
   },
-  "version": "0.8.27"
+  "version": "0.8.17"
 }
