+++ examples/upload/package.json
@@ -1,6 +1,6 @@
 {
   "name": "upload",
-  "version": "0.8.27",
+  "version": "0.8.17",
   "description": "Demonstration of uploading an image to UpscalerJS",
   "main": "index.js",
   "scripts": {
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@tensorflow/tfjs": "2.8.6",
-    "upscaler": "0.8.27"
+    "upscaler": "0.8.17"
   },
   "browserslist": {
     "production": [
