+++ examples/warmup/package.json
@@ -1,6 +1,6 @@
 {
   "name": "warmup",
-  "version": "0.8.27",
+  "version": "0.8.17",
   "description": "Demonstration of warming up UpscalerJS",
   "main": "index.js",
   "scripts": {
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@tensorflow/tfjs": "2.8.6",
-    "upscaler": "0.8.27"
+    "upscaler": "0.8.17"
   },
   "browserslist": {
     "production": [
