+++ packages/upscalerjs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "upscaler",
-  "version": "0.8.27",
+  "version": "0.8.17",
   "description": "Increase Image resolution with Tensorflow.js",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
