+++ scripts/pin-cdn.js
@@ -1,9 +1,6 @@
 /*****
- * This script does two things:
- * 
- * - pins the current version of UpscalerJS to the correct CDN path, so that version x.x.x of upscaler
+ * Pins the current version of UpscalerJS to the correct CDN path, so that version x.x.x of upscaler
  *      points to the same version of @upscaler/models
- * - updates the upscaler dependencies in each examples folder to the current version
  */
 const fs = require('fs');
 const path = require('path');
@@ -11,13 +8,8 @@ const path = require('path');
 const ROOT = path.resolve(__dirname, '..');
 const UPSCALERJS_ROOT = path.resolve(ROOT, 'packages/upscalerjs');
 
-const ROOT_PACKAGE = require(path.resolve(ROOT, 'package.json'));
-const UPSCALERJS_PACKAGE = require(path.resolve(ROOT, 'packages/upscalerjs/package.json'));
 const MODELS_PACKAGE = require(path.resolve(ROOT, 'packages/models/package.json'));
 
-const peerDependencies = ROOT_PACKAGE['peerDependencies'];
-const version = ROOT_PACKAGE['version'];
-const upscalerVersion = UPSCALERJS_PACKAGE['version'];
 const modelsVersion = MODELS_PACKAGE['version'];
 
 console.log(`Pinning version ${modelsVersion} in CDN definition`);
@@ -33,30 +25,3 @@ const writeConstantsFile = (version) => {
   );
 }
 writeConstantsFile(modelsVersion);
-
-const updatePackageJSON = (packagePath, version, callback) => {
-  const package = require(packagePath);
-  package.version = version;
-  if (callback) {
-    callback(package);
-  }
-  fs.writeFileSync(packagePath, JSON.stringify(package, null, 2));
-  console.log(`Wrote file ${packagePath}`);
-}
-
-const examplesPath = path.resolve(ROOT, 'examples');
-const examples = fs.readdirSync(examplesPath)
-examples.forEach(example => {
-  const dirPath = path.resolve(examplesPath, example);
-  if (fs.lstatSync(dirPath).isDirectory()) {
-    const examplePackagePath = path.resolve(dirPath, 'package.json');
-    updatePackageJSON(examplePackagePath, upscalerVersion, package => {
-      if (!package.dependencies) {
-        package.dependencies = {};
-      }
-      package.dependencies.upscaler = upscalerVersion;
-      
-      package.dependencies['@tensorflow/tfjs'] = peerDependencies['@tensorflow/tfjs'];
-    });
-  }
-});
