+++ scripts/start_example.js
